+++ NEWS.md
@@ -1,8 +1,9 @@
 # DataPackageR 0.15.3
-* conditional tests when pandoc is missing
+* conditional tests when pandoc is missing (ropensci/DataPackager/issues/46)
+* add use_data_object and use_processing_script (ropensci/DataPackager/issues/44)
 
 # DataPackageR 0.15.2
-* Add  pandoc to SystemRequirements
+* Add  pandoc to SystemRequirements (ropensci/DataPackager/issues/46)
 * Add use_raw_dataset() method (and tests) to add data sets to inst/extdata. interactively. (ropensci/DataPackager/issues/44)
 
 # DataPackageR 0.15.1.9000